Add conditionals for comparison
each one will throw a message and return a variable. This will be used in game to determine
who gets a point added to their score. Should the value return 0, neither conditionals in game
will fire, and the value will be reset to 0 at the end of the for loop.