Add event listener for each button and class for each button
Added an event listener in game() in js file. Uses .querySelectorAll to select each button.
Then a forEach call, which listens for each individual click. The buttons themselves
recieved class designations within the index.html file. When each button is clicked
a temporary alert message is sent with the className of the item clicked.