Move score items to scoreBoard function
Moved the player, computer, and tie variables to global values, as to not lose their
value in memory when moving outside the scope of scoreBoard. Changed scoreBoards accepted
value to round, and create a switch case in the function that increments the global values.
This was done as the intention is to lock out all buttons except reset upon reaching a score
of 5. The intention is to reset the global values should the player wish to play again, delete
the reset button and increment again to 5.